+++ ocaml-lsp-server/src/code_actions/action_open_related.mli
@@ -2,6 +2,8 @@ open Import
 
 val command_name : string
 
+val available : ShowDocumentClientCapabilities.t option -> bool
+
 val command_run : _ Server.t -> ExecuteCommandParams.t -> Json.t Fiber.t
 
 val for_uri :