+++ CHANGES.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Fixes
+
+- Respect `showDocument` capabilities. Do not offer commands or code actions
+  that rely on this request without client support. (#836)
+
 # 1.13.1
 
 ## Fixes